Switched to branch longer

+++ src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
 
     public static void main(String[] args) {
-        int[] sales = new int[] {100, 200, 300, 400};
+        long[] sales = new long[] {100, 200, 300, 400};
         SalesManager salesManager = new SalesManager(sales);
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println(max);
     }
 }
